# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/user/login":{"post":{"tags":["User"],"summary":"User Login","operationId":"user_login_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user":{"get":{"tags":["User"],"summary":"Get Current User","operationId":"get_current_user_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserData"}}}}},"security":[{"JWTBearer":[]}]},"post":{"tags":["User"],"summary":"User Signup","operationId":"user_signup_user_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserData"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}},"/note":{"get":{"tags":["Note"],"summary":"Get Note","operationId":"get_note_note_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"JWTBearer":[]}]},"post":{"tags":["Note"],"summary":"Update Note","operationId":"update_note_note_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Note"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Note":{"title":"Note","required":["note"],"type":"object","properties":{"note":{"title":"Note","type":"string"}},"example":{"note":"Note Text"}},"User":{"title":"User","required":["fullname","email","password"],"type":"object","properties":{"fullname":{"title":"Fullname","minLength":1,"type":"string"},"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","minLength":1,"type":"string"}},"example":{"fullname":"Anmol Garg","email":"anmol@test.com","password":"qwerty1234"}},"UserData":{"title":"UserData","required":["fullname","email","is_super_admin"],"type":"object","properties":{"fullname":{"title":"Fullname","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"is_super_admin":{"title":"Is Super Admin","type":"boolean"}},"example":{"fullname":"Anmol Garg","email":"qwerty1234","is_super_admin":false}},"UserSignIn":{"title":"UserSignIn","required":["username","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"}},"example":{"username":"anmol@test.com","password":"qwerty1234"}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"JWTBearer":{"type":"http","scheme":"bearer"}}}}